fix: when using insecure connection the PROXY_HTTPS env var was ignored (#12)
* Read proxy envvar when setting insecure connection

When the insecure parameter is set, the transport is overriden, the
commit add the proxy configuration from environnement variable.

* Use default transport with insecure connection

Instead of having a custom blank transport when insecure connection is
set, clone the default transport and disable TLS verification. This
allow to have standard behaviour, like reading HTTP_PROXY and NO_PROXY
variables.